[compiler-rt][purecap] Fix more typing issues (7)

+++ compiler-rt/lib/asan/asan_allocator.cpp
@@ -182,7 +182,7 @@ class LargeChunkHeader {
       return;
     }
 
-    size_t old = kAllocBegMagic;
+    unsigned long long old = kAllocBegMagic;
     if (!atomic_compare_exchange_strong(&magic, &old, 0,
                                         memory_order_release)) {
       CHECK_EQ(old, kAllocBegMagic);
